extract zero
